+++ menu/fileSelector.hpp
@@ -176,7 +176,7 @@ struct FileSelector
 				for (std::size_t i = 0; i < p->children.size(); ++i)
 				{
 					if (ch == p->children[i].get())
-						p->getMenu().moveTo(i);
+						p->getMenu().moveTo((int)i);
 				}
 			}
